renamed example_project to example_app
It's an app not a project, so was misleadingly named
