+++ piccolo/query/methods/select.py
@@ -19,8 +19,8 @@ from piccolo.query.mixins import (
 from piccolo.querystring import QueryString
 
 if t.TYPE_CHECKING:
-    from piccolo.table import Table  # noqa
     from piccolo.custom_types import Combinable
+    from piccolo.table import Table  # noqa
 
 
 class Count(Selectable):
