+++ tests/engine/test_transaction.py
@@ -1,7 +1,7 @@
 import asyncio
 from unittest import TestCase
 
-from ..example_project.tables import Band, Manager
+from ..example_app.tables import Band, Manager
 from ..base import postgres_only
