+++ tests/example_app/piccolo_app.py
@@ -1,0 +1,6 @@
+from piccolo.conf.apps import AppConfig
+
+
+APP_CONFIG = AppConfig(
+    app_name="example_app", migrations_folder_path="", commands=[]
+)
