+++ tests/query/test_await.py
@@ -1,7 +1,7 @@
 import asyncio
 
 from ..base import DBTestCase
-from ..example_project.tables import Band
+from ..example_app.tables import Band
 
 
 class TestAwait(DBTestCase):
