+++ tests/sqlite_conf.py
@@ -5,4 +5,5 @@ from piccolo.conf.apps import AppRegistry
 DB = SQLiteEngine(path="test.sqlite")
 
 
-APP_REGISTRY = AppRegistry()
+APP_REGISTRY = AppRegistry(apps=["tests.example_app.piccolo_app"])
+
