+++ tests/table/instance/test_get_related.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from tests.example_project.tables import Manager, Band
+from tests.example_app.tables import Manager, Band
 
 
 TABLES = [Manager, Band]
