+++ tests/table/test_indexes.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
 from ..base import DBTestCase
-from ..example_project.tables import Manager
+from ..example_app.tables import Manager
 
 
 class TestIndexes(DBTestCase):
