+++ tests/table/test_metaclass.py
@@ -1,9 +1,9 @@
 from unittest import TestCase
 
-from ..example_project.tables import Band
+from ..example_app.tables import Band
 
 
 class TestMetaClass(TestCase):
-
     def test_tablename(self):
-        self.assertEqual(Band._meta.tablename, 'band')
+        self.assertEqual(Band._meta.tablename, "band")
+
