+++ tests/table/test_output.py
@@ -1,7 +1,7 @@
 import json
 
 from ..base import DBTestCase
-from ..example_project.tables import Band
+from ..example_app.tables import Band
 
 
 class TestOutputList(DBTestCase):
