+++ tests/table/test_repr.py
@@ -1,5 +1,5 @@
 from ..base import DBTestCase
-from ..example_project.tables import Manager
+from ..example_app.tables import Manager
 
 
 class TestTableRepr(DBTestCase):
